Use SASS to move stylesheets into the app/ directory, and make them all .scss files. SASS will conveniently process and place the files in the app's public/stylesheets directory as needed, removing the requirment to copy them over when installing the CMS gem. (This is good as prep-work for Rails 3.1).
Minimal SCSS syntax is used--to encapsulate all styling to within the #cms_body_wrapper for better integration with any application layout. Extracted styles the are applied to html and body tags into a separate file: this is not likely going to be used if the application has its own layout.