Updated the French locale to sound more natural
This is the first time I do a pull request.

I've done my best, but there are some places which can't be translated like "Published" or not well like "Revisions for" which varies in French depending on the word following.

Translations of label, slug and identifier could certainly be improved.