Split out the functionality of cms_page_content.
cms_page_content was very slow when due to having to determine the type
of content it was tasked with retrieving. Furthermore, the method would
behave differently based on what content was being fetched.

Since the method method behaves differently based on the content being
fetched, it was required that the caller know what type of content it
was expecting. As such, it wasn't much of a stretch to split the method
into two distinct methods.

Change-Id: I342381ff863e17529659dba7ba9148d81368b18d