Add processed content field when saving page content
The Comfy and CMS has a property on the page that is
called "content_cache" but it is a default from Comfy
and does not use the Mastalk gem and uses some not
organised code on the lib directory.
So I decide to add a columns processed_content
that solves the issue.