Add precompiled assets explicitly (#840)
* Add precompiled assets explicitly

This should be explicit because:

a) The user may override the default `config.assets.precompile`, removing a default `/application\.(css|js)/` entry.

b) The user may also be using Sprockets 4, which only has `manifest.js` in `config.assets.precompile` by default. Not sure if this alone is enough for Sprockets 4 compatibility though.