Write label attribute before save using filename
Problem was that when uploading a new file the actual
label would be empty. Model will return a label when
requested as it will read the file name if label is
empty, but when doing in rails helpers something like

Comfy::Cms::File.where(label: "name")

the file will not be found.

This fix will user before_save callback to write
the label attribute on the model using the file name
which will be saved then in DB


Signed-off-by: oscar sanhueza <oscar@hashdot.fi>