Update label in db when process attaching the file
Problem was that when uploading a new file the actual
label would be empty. Model will return a label when
requested as it will read the file name if label is
empty, but when doing in rails helpers something like

Comfy::Cms::File.where(label: "name")

the file will not be found.

This fix will call update_attribute on the label
after the file is save in controller and the actual
file is attached to the model, simple enough.


Signed-off-by: oscar sanhueza <oscar@hashdot.fi>